Refactor ProposalBuilder and Proposal Creation Pages to Enhance Actions Experience
- Removed unused imports and components from the ProposalBuilder, streamlining the code.
- Introduced a new `ActionsExperience` component to encapsulate action handling logic, improving modularity and readability.
- Updated the `SafeProposalWithActionsCreatePage` to utilize the new `ActionsExperience` component, enhancing the user interface for action management.
- Set `actionsExperience` prop to `null` in other proposal creation pages to maintain consistency where actions are not applicable.

These changes aim to improve the organization and clarity of the proposal creation process, particularly in managing proposal actions.
