+++ src/pages/dao/proposals/actions/new/SafeProposalWithActionsCreatePage.tsx
@@ -1,5 +1,5 @@
 import * as amplitude from '@amplitude/analytics-browser';
-import { Center } from '@chakra-ui/react';
+import { Center, Flex, Text } from '@chakra-ui/react';
 import { useEffect } from 'react';
 import { useTranslation } from 'react-i18next';
 import { Route, useLocation, useNavigate } from 'react-router-dom';
@@ -7,10 +7,13 @@ import {
   ProposalBuilder,
   ShowNonceInputOnMultisig,
 } from '../../../../../components/ProposalBuilder';
+import { ProposalActionCard } from '../../../../../components/ProposalBuilder/ProposalActionCard';
 import { DEFAULT_PROPOSAL_METADATA_TYPE_PROPS } from '../../../../../components/ProposalBuilder/ProposalMetadata';
 import ProposalTransactionsForm from '../../../../../components/ProposalBuilder/ProposalTransactionsForm';
 import { DEFAULT_PROPOSAL } from '../../../../../components/ProposalBuilder/constants';
 import { BarLoader } from '../../../../../components/ui/loaders/BarLoader';
+import { AddActions } from '../../../../../components/ui/modals/AddActions';
+import { SendAssetsData } from '../../../../../components/ui/modals/SendAssetsModal';
 import { useHeaderHeight } from '../../../../../constants/common';
 import { DAO_ROUTES } from '../../../../../constants/routes';
 import { usePrepareProposal } from '../../../../../hooks/DAO/proposal/usePrepareProposal';
@@ -19,7 +22,66 @@ import { useFractal } from '../../../../../providers/App/AppProvider';
 import { useNetworkConfigStore } from '../../../../../providers/NetworkConfig/useNetworkConfigStore';
 import { useProposalActionsStore } from '../../../../../store/actions/useProposalActionsStore';
 import { useDaoInfoStore } from '../../../../../store/daoInfo/useDaoInfoStore';
-import { CreateProposalSteps, ProposalBuilderMode } from '../../../../../types';
+import { CreateProposalSteps, ProposalActionType, ProposalBuilderMode } from '../../../../../types';
+
+function ActionsExperience() {
+  const { t } = useTranslation('actions');
+  const { actions, addAction } = useProposalActionsStore();
+
+  const handleAddSendAssetsAction = (data: SendAssetsData) => {
+    addAction({
+      actionType: ProposalActionType.TRANSFER,
+      content: <></>,
+      transactions: [
+        {
+          targetAddress: data.asset.tokenAddress,
+          ethValue: {
+            bigintValue: 0n,
+            value: '0',
+          },
+          functionName: 'transfer',
+          parameters: [
+            { signature: 'address', value: data.destinationAddress },
+            { signature: 'uint256', value: data.transferAmount.toString() },
+          ],
+        },
+      ],
+    });
+  };
+
+  return (
+    <Flex
+      flexDirection="column"
+      gap="1.5rem"
+    >
+      <Flex
+        flexDirection="column"
+        gap="0.5rem"
+      >
+        <Flex
+          mt={6}
+          mb={2}
+          alignItems="center"
+        >
+          <Text ml={2}>{t('actions', { ns: 'actions' })}</Text>
+        </Flex>
+        {actions.map((action, index) => {
+          return (
+            <ProposalActionCard
+              key={index}
+              action={action}
+              index={index}
+              canBeDeleted={actions.length > 1}
+            />
+          );
+        })}
+      </Flex>
+      <Flex>
+        <AddActions addSendAssetsAction={handleAddSendAssetsAction} />
+      </Flex>
+    </Flex>
+  );
+}
 
 export function SafeProposalWithActionsCreatePage() {
   useEffect(() => {
@@ -78,6 +140,7 @@ export function SafeProposalWithActionsCreatePage() {
       pageHeaderTitle={t('createProposal', { ns: 'proposal' })}
       pageHeaderBreadcrumbs={pageHeaderBreadcrumbs}
       pageHeaderButtonClickHandler={pageHeaderButtonClickHandler}
+      actionsExperience={<ActionsExperience />}
       proposalMetadataTypeProps={DEFAULT_PROPOSAL_METADATA_TYPE_PROPS(t)}
       prevStepUrl={prevStepUrl}
       nextStepUrl={nextStepUrl}
