+++ src/pages/dao/proposals/new/sablier/SafeSablierProposalCreatePage.tsx
@@ -149,6 +149,7 @@ export function SafeSablierProposalCreatePage() {
       pageHeaderBreadcrumbs={pageHeaderBreadcrumbs}
       pageHeaderButtonClickHandler={pageHeaderButtonClickHandler}
       proposalMetadataTypeProps={DEFAULT_PROPOSAL_METADATA_TYPE_PROPS(t)}
+      actionsExperience={null}
       prevStepUrl={prevStepUrl}
       nextStepUrl={nextStepUrl}
       prepareProposalData={prepareProposalData}
